todo: fix edit & delete grocery list
i tried fixing it and added some stuff to update household & user but could not get it to work

+++ data/groceryList.js
@@ -1,23 +1,26 @@
 import { users, groceryLists, household } from "../config/mongoCollections.js";
 import { checkHouseholdName, checkId, checkString } from '../validation.js';
 import userData from './users.js'
-import {ObjectId} from 'mongodb';
+import { ObjectId } from 'mongodb';
 const exportedMethods = {
-    async newGroceryList (
+  async newGroceryList(
     userId,
     householdName,
     groceryName,
     listType
   ) {
+    if (!userId || !groceryName || !householdName || !listType) {
+      throw 'Must provide all fields to update Grocery List';
+    }
 
     userId = checkId(userId, "User ID");
     householdName = checkHouseholdName(householdName, "Household Name");
     groceryName = checkString(groceryName, "Grocery List Name");
-    if(listType.toLowerCase() !== 'community')
-     if (listType.toLowerCase() !== 'special occasion')
-      if (listType.toLowerCase() !== 'personal') {
-      throw 'Not a valid list type';
-    }
+    if (listType.toLowerCase() !== 'community')
+      if (listType.toLowerCase() !== 'special occasion')
+        if (listType.toLowerCase() !== 'personal') {
+          throw 'Not a valid list type';
+        }
     const userMember = await userData.getUserById(userId);
     if (!userMember) throw `User could not be found`;
 
@@ -58,7 +61,7 @@ const exportedMethods = {
     // update household with new grocery list
     let updatedInfo = await householdCollection.findOneAndUpdate(
       { _id: new ObjectId(existingHousehold[0]._id) },
-      { $push: {groceryLists: newList._id.toString()}}, // push ID of list to household
+      { $push: { groceryLists: newList._id.toString() } }, // push ID of list to household
       { returnDocument: 'after' }
     );
     if (!updatedInfo) throw 'Error: Could not update household successfully';
@@ -67,12 +70,12 @@ const exportedMethods = {
     const userCollection = await users();
     updatedInfo = await userCollection.findOneAndUpdate(
       { _id: userMember._id },
-      { $push: {groceryLists: newList._id.toString()}}, // push ID of list to household
+      { $push: { groceryLists: newList._id.toString() } }, // push ID of list to household
       { returnDocument: 'after' }
     );
     return newList;
   },
-  
+
   async getAllGroceryLists() {
     const groceryListCollection = await groceryLists();
     let gList = await groceryListCollection.find({}).toArray();
@@ -85,59 +88,74 @@ const exportedMethods = {
 
   async getGroceryList(id) {
     id = checkId(id, "Grocery List Id");
-    const groceryListCollection = await groceryLists(); 
+    const groceryListCollection = await groceryLists();
     const gList = await groceryListCollection.findOne({ _id: new ObjectId(id) });
-    if (!gList){
+    if (!gList) {
       throw 'Error: List not found.'
     }
     return gList;
   },
 
-  async deleteGroceryList(id) {
+  async deleteGroceryList(id, householdName) {
     id = checkId(id, "Grocery List Id");
-    const groceryListCollection = await groceryLists(); 
+    householdName = checkHouseholdName(householdName, "Household Name");
+
+    let groceryList = await this.getGroceryList(id);
+    const householdCollection = await household();
+    const existingHousehold = await householdCollection.find({ householdName: householdName }).toArray();
+    if (existingHousehold.length === 0) throw `Error: Household does not exist`;
+    let updatedInfo = await householdCollection.findOneAndUpdate(
+      { _id: new ObjectId(existingHousehold[0]._id) },
+      { $pull: { groceryLists: groceryList._id.toString() } }, // push ID of list from household
+      { returnDocument: 'after' }
+    );
+    if (!updatedInfo) throw 'Error: Could not update household successfully';
+
+    // update user to remove grocery list
+    const userCollection = await users();
+    updatedInfo = await userCollection.findOneAndUpdate(
+      { _id: groceryList.userId },
+      { $pull: { groceryLists: groceryList._id.toString() } }, // pull ID of list from user
+      { returnDocument: 'after' }
+    );
+    if (!updatedInfo) throw 'Error: Could not update user successfully';
+
+    const groceryListCollection = await groceryLists();
     const deleteInfo = await groceryListCollection.findOneAndDelete({ _id: new ObjectId(id) });
-    if (!deleteInfo){
-      throw 'Error: Deletion unsuccessful'; 
+    if (!deleteInfo) {
+      throw 'Error: Deletion unsuccessful';
     }
 
-    // TODO: update user & pull grocery list from user's grocery lists
-
-    return {groceryListDeleted: true};
+    return { groceryListDeleted: true };
   },
-  
-  async updateGroceryList(
-    userId,
-    groceryName,
-    items,
-  ) {
-    if(!userId || !groceryName || !items){
-      throw 'Must provide all fields to update Grocery List';
-    }
 
-    userId = checkId(userId, "User Id");
+  async updateGroceryList(
+    id,
+    groceryName,
+  ) {
+//    should we make fields optional for update?
+//    if (!userId || !groceryName || !listType) {
+//      throw 'Must provide all fields to update Grocery List';
+//    }
+
+    id = checkId(id, "Grocery List Id");
     groceryName = checkString(groceryName, "Grocery List Name");
 
     let name = groceryName.trim();
-    if(name.length === 0){
+    if (name.length === 0) {
       throw 'You must provide an input for the list name';
     }
 
-    const userMember = await userData.getUserById(userId);
-
     const updateGroceryList = {
-      userName: userMember.firstName + " " + userMember.lastName,
       groceryName: name,
-      items: items,
     }
 
     const groceryListCollection = await groceryLists();
-    const gList = await groceryListCollection.findOneAndUpdate({ _id: new ObjectId(id) }, {$set: updateGroceryList}, { returnOriginal: false } );
-    if (!gList){
-      throw 'Error: Could not update list'
-    }
-    return {groceryListUpdated: true};
-    }
+    const gList = await groceryListCollection.findOneAndUpdate({ _id: new ObjectId(id) }, { $set: updateGroceryList }, { returnDocument: 'after' });
+    if (!gList) throw 'Error: Could not update list'
 
-  };
-  export default exportedMethods;
+    return updateGroceryList;
+  }
+
+};
+export default exportedMethods;
